+++ HISTORY.rst
@@ -15,3 +15,5 @@ History
 * Allow users to change the sharpness of the regional border lines by channging the DPI value used when creating the masks.
 * Add wordcloud colour function based on rank of words.
 * Add new function ``calc_tfidf()`` to calculate TF-IDF score of each word in each region in a dataframe.
+* Allow users to use shapefiles form a local filepath instead of downloading from GISCO's online database.
+* Change default coordination system when downloading shapefiles.